+++ packages/block-library/src/button/index.js
@@ -38,7 +38,7 @@ export const settings = {
 	} ),
 };
 
-if ( window.__experimentalContentOnlyPatternInsertion ) {
+if ( window.__experimentalContentOnlyInspectorFields ) {
 	settings[ fieldsKey ] = [
 		{
 			id: 'text',