+++ packages/global-styles-ui/src/font-library-modal/context.tsx
@@ -2,7 +2,12 @@
  * WordPress dependencies
  */
 import { createContext, useState, useEffect } from '@wordpress/element';
-import { useSelect, useDispatch } from '@wordpress/data';
+import {
+	useSelect,
+	useDispatch,
+	resolveSelect,
+	useRegistry,
+} from '@wordpress/data';
 import {
 	useEntityRecord,
 	useEntityRecords,
@@ -13,15 +18,17 @@ import type {
 	FontFamilyPreset,
 	GlobalStylesConfig,
 } from '@wordpress/global-styles-engine';
+import type {
+	CollectionFontFace,
+	CollectionFontFamily,
+	FontFace,
+	FontFamily,
+} from '@wordpress/core-data';
 
 /**
  * Internal dependencies
  */
-import {
-	fetchGetFontFamilyBySlug,
-	fetchInstallFontFamily,
-	fetchUninstallFontFamily,
-} from './resolvers';
+import { fetchInstallFontFamily } from './api';
 import {
 	setUIValuesNeeded,
 	mergeFontFamilies,
@@ -35,12 +42,7 @@ import {
 } from './utils';
 import { setImmutably } from './utils/set-immutably';
 import { toggleFont } from './utils/toggleFont';
-import type {
-	CollectionFontFamily,
-	FontFace,
-	FontFamily,
-	FontLibraryState,
-} from './types';
+import type { FontFamilyToUpload, FontLibraryState } from './types';
 import { useSetting } from '../hooks';
 
 export const FontLibraryContext = createContext< FontLibraryState >(
@@ -49,7 +51,8 @@ export const FontLibraryContext = createContext< FontLibraryState >(
 FontLibraryContext.displayName = 'FontLibraryContext';
 
 function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
-	const { saveEntityRecord } = useDispatch( coreStore );
+	const registry = useRegistry();
+	const { saveEntityRecord, deleteEntityRecord } = useDispatch( coreStore );
 	const { globalStylesId } = useSelect( ( select ) => {
 		const { __experimentalGetCurrentGlobalStylesId } = select( coreStore );
 		return { globalStylesId: __experimentalGetCurrentGlobalStylesId() };
@@ -62,18 +65,12 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 	);
 
 	const [ isInstalling, setIsInstalling ] = useState( false );
-	const [ refreshKey, setRefreshKey ] = useState( 0 );
-
-	const refreshLibrary = () => {
-		setRefreshKey( Date.now() );
-	};
 
 	const { records: libraryPosts = [], isResolving: isResolvingLibrary } =
 		useEntityRecords< CollectionFontFamily >(
 			'postType',
 			'wp_font_family',
 			{
-				refreshKey,
 				_embed: true,
 			}
 		);
@@ -112,6 +109,9 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 			| Record< string, FontFamilyPreset[] >
 			| undefined
 	) => {
+		if ( ! globalStyles.record ) {
+			return;
+		}
 		// Gets the global styles database post content.
 		const updatedGlobalStyles = globalStyles.record;
 
@@ -230,7 +230,7 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 		return getActivatedFontsOutline( source )[ slug ] || [];
 	};
 
-	async function installFonts( fontFamiliesToInstall: FontFamily[] ) {
+	async function installFonts( fontFamiliesToInstall: FontFamilyToUpload[] ) {
 		setIsInstalling( true );
 		try {
 			const fontFamiliesToActivate = [];
@@ -242,16 +242,40 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 				let isANewFontFamily = false;
 
 				// Get the font family if it already exists.
-				let installedFontFamily = await fetchGetFontFamilyBySlug(
-					fontFamilyToInstall.slug
-				);
+				const fontFamilyRecords = await resolveSelect(
+					coreStore
+				).getEntityRecords( 'postType', 'wp_font_family', {
+					slug: fontFamilyToInstall.slug,
+					per_page: 1,
+					_embed: true,
+				} );
+
+				const fontFamilyPost =
+					fontFamilyRecords && fontFamilyRecords.length > 0
+						? fontFamilyRecords[ 0 ]
+						: null;
+
+				let installedFontFamily = fontFamilyPost
+					? {
+							id: fontFamilyPost.id,
+							...fontFamilyPost.font_family_settings,
+							fontFace:
+								(
+									fontFamilyPost?._embedded?.font_faces ?? []
+								).map(
+									( face: CollectionFontFace ) =>
+										face.font_face_settings
+								) || [],
+					  }
+					: null;
 
 				// Otherwise create it.
 				if ( ! installedFontFamily ) {
 					isANewFontFamily = true;
 					// Prepare font family form data to install.
 					installedFontFamily = await fetchInstallFontFamily(
-						makeFontFamilyFormData( fontFamilyToInstall )
+						makeFontFamilyFormData( fontFamilyToInstall ),
+						registry
 					);
 				}
 
@@ -294,7 +318,8 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 						installedFontFamily.id,
 						makeFontFacesFormData(
 							fontFamilyToInstall as FontFamily
-						)
+						),
+						registry
 					);
 					successfullyInstalledFontFaces = response?.successes;
 					unsuccessfullyInstalledFontFaces = response?.errors;
@@ -329,7 +354,12 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 					( fontFamilyToInstall?.fontFace?.length ?? 0 ) > 0 &&
 					successfullyInstalledFontFaces?.length === 0
 				) {
-					await fetchUninstallFontFamily( installedFontFamily.id );
+					await deleteEntityRecord(
+						'postType',
+						'wp_font_family',
+						installedFontFamily.id,
+						{ force: true }
+					);
 				}
 
 				installationErrors = installationErrors.concat(
@@ -353,8 +383,6 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 				);
 				// Save the global styles to the database.
 				await saveFontFamilies( activeFonts );
-
-				refreshLibrary();
 			}
 
 			if ( installationErrorMessages.length > 0 ) {
@@ -378,24 +406,18 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 		try {
 			// Uninstall the font family.
 			// (Removes the font files from the server and the posts from the database).
-			const uninstalledFontFamily = await fetchUninstallFontFamily(
-				fontFamilyToUninstall.id
+			await deleteEntityRecord(
+				'postType',
+				'wp_font_family',
+				fontFamilyToUninstall.id,
+				{ force: true }
 			);
 
-			// Deactivate the font family if delete request is successful
-			// (Removes the font family from the global styles).
-			if ( uninstalledFontFamily.deleted ) {
-				const activeFonts = deactivateFontFamily(
-					fontFamilyToUninstall
-				);
-				// Save the global styles to the database.
-				await saveFontFamilies( activeFonts );
-			}
-
-			// Refresh the library (the library font families from database).
-			refreshLibrary();
-
-			return uninstalledFontFamily;
+			// Deactivate the font family (remove from global styles).
+			const activeFonts = deactivateFontFamily( fontFamilyToUninstall );
+			// Save the global styles to the database.
+			await saveFontFamilies( activeFonts );
+			return { deleted: true };
 		} catch ( error ) {
 			// eslint-disable-next-line no-console
 			console.error(
@@ -538,7 +560,6 @@ function FontLibraryProvider( { children }: { children: React.ReactNode } ) {
 				getAvailableFontsOutline,
 				modalTabOpen,
 				setModalTabOpen,
-				refreshLibrary,
 				saveFontFamilies,
 				isResolvingLibrary,
 				isInstalling,