+++ packages/global-styles-ui/src/font-library-modal/utils/toggleFont.ts
@@ -31,9 +31,9 @@
  * // This will add the specified face to 'roboto' in customFonts
  */
 /**
- * Internal dependencies
+ * WordPress dependencies
  */
-import type { FontFamily, FontFace } from '../types';
+import type { FontFamily, FontFace } from '@wordpress/core-data';
 
 export function toggleFont(
 	font: FontFamily,